requirements.txt
Django==4.2.14
Pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
dj-database-url==2.1.0
psycopg2-binary==2.9.9
supabase==2.4.0
gunicorn==22.0.0
whitenoise==6.7.0
